+++ app/src/main/java/com/dogandpigs/fitnote/presentation/lesson/memberlessonlist/MemberLessonListViewModel.kt
@@ -30,7 +30,7 @@ internal class MemberLessonListViewModel @Inject constructor(
                 memberRepository.getMemberList()
             }.onSuccess {
                 val member = it?.memberList?.first { member ->
-                    member.id == memberId.toLong()
+                    member.id == memberId
                 }
 
                 member?.also {
